pyproject.toml
[project]
name = "dl"
version = "0.1.0"
description = ""
authors = [
    {name = "trclarsi",email = "tr.clarsi@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.15"
dependencies = [
    "pandas (>=2.3.3,<3.0.0)",
    "matplotlib (>=3.10.7,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "numpy (>=2.3.4,<3.0.0)",
    "ipykernel (>=7.1.0,<8.0.0)",
    "jupyterlab (>=4.4.10,<5.0.0)",
    "tensorflow (>=2.20.0,<3.0.0)",
    "scipy (>=1.16.3,<2.0.0)",
    "jupyter (>=1.1.1,<2.0.0)",
    "docling (>=2.63.0,<3.0.0)",
    "faiss-cpu (>=1.13.0,<2.0.0)",
    "mistralai (==0.4.2)",
    "openai-whisper (>=20250625,<20250626)",
    "gensim (>=4.4.0,<5.0.0)",
    "tf-keras (>=2.20.1,<3.0.0)",
    "mistral (>=21.0.0,<22.0.0)",
    "google-generativeai (>=0.8.6,<0.9.0)",
    "ragas (>=0.4.2,<0.5.0)",
    "langchain-google-genai (>=4.1.2,<5.0.0)",
    "datasets (>=4.4.2,<5.0.0)",
    "groq (>=0.9.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
